+++ Hardware/NMEADevice.cpp
@@ -22,7 +22,7 @@ THREAD_PROC_RETURN_VALUE NMEADeviceThread(void* pParam)
 	CHRONO chrono_period;
 	int threadperiod = 100;
 	int errcount = 0;
-	int deviceid = (intptr_t)pParam;
+	int deviceid = (int)(intptr_t)pParam;
 	char szCfgFilePath[256];
 	int i = 0;
 	char szSaveFilePath[256];
@@ -94,7 +94,7 @@ THREAD_PROC_RETURN_VALUE NMEADeviceThread(void* pParam)
 						sprintf(szTemp, "nmeadevice");
 					}
 					// Remove the extension.
-					for (i = strlen(szTemp)-1; i >= 0; i--) { if (szTemp[i] == '.') break; }
+					for (i = (int)strlen(szTemp)-1; i >= 0; i--) { if (szTemp[i] == '.') break; }
 					if ((i > 0)&&(i < (int)strlen(szTemp))) memset(szTemp+i, 0, strlen(szTemp)-i);
 					//if (strlen(szTemp) > 4) memset(szTemp+strlen(szTemp)-4, 0, 4);
 					EnterCriticalSection(&strtimeCS);
